stories inteface
