+++ packages/portal/src/components/stories/StoriesInterface.vue
@@ -1,0 +1,219 @@
+<template>
+  <div
+    data-qa="stories interface"
+  >
+    <StoriesTagsDropdown
+      :filtered-tags="filteredTags"
+      :selected-tags="selectedTags"
+    />
+    <div
+      class="mb-4 context-label"
+    >
+      {{ $tc('items.itemCount', total, { count: total }) }}
+    </div>
+    <b-container
+      v-if="loadingStories"
+      data-qa="stories loading spinner container"
+    >
+      <b-row class="flex-md-row py-4 text-center">
+        <b-col cols="12">
+          <LoadingSpinner />
+        </b-col>
+      </b-row>
+    </b-container>
+    <b-card-group
+      v-else
+      class="card-deck-4-cols gridless-browse-cards"
+      deck
+    >
+      <template
+        v-for="(entry, index) in stories"
+      >
+        <template
+          v-if="entry === ctaBanner"
+        >
+          <CallToActionBanner
+            v-if="callToAction"
+            :key="index"
+            :name="callToAction.name"
+            :text="callToAction.text"
+            :link="callToAction.relatedLink"
+            :illustration="callToAction.image"
+            class="cta-banner"
+          />
+        </template>
+        <ContentCard
+          v-else-if="entry !== ctaBanner"
+          :key="index"
+          :title="entry.name"
+          :url="entryUrl(entry)"
+          :image-url="entry.primaryImageOfPage && entry.primaryImageOfPage.image.url"
+          :image-content-type="entry.primaryImageOfPage && entry.primaryImageOfPage.image.contentType"
+          :image-optimisation-options="entry.primaryImageOfPage ? entryImageOptions(entry.primaryImageOfPage.image) : {}"
+        />
+      </template>
+    </b-card-group>
+    <PaginationNavInput
+      v-if="total > perPage"
+      :per-page="perPage"
+      :total-results="total"
+    />
+  </div>
+</template>
+
+<script>
+  import uniq from 'lodash/uniq';
+  import StoriesTagsDropdown from '@/components/stories/StoriesTagsDropdown';
+  import ContentCard from '@/components/generic/ContentCard';
+  import LoadingSpinner from '@/components/generic/LoadingSpinner';
+
+  const CTA_BANNER = 'cta-banner';
+
+  export default {
+    name: 'StoriesInterface',
+
+    components: {
+      CallToActionBanner: () => import('@/components/generic/CallToActionBanner'),
+      StoriesTagsDropdown,
+      ContentCard: () => import('@/components/generic/ContentCard'),
+      LoadingSpinner: () => import('../generic/LoadingSpinner'),
+      PaginationNavInput: () => import('../generic/PaginationNavInput')
+    },
+
+
+    props: {
+      callToAction: {
+        type: Object,
+        default: () => {}
+      }
+    },
+
+    data() {
+      return {
+        allStoryMetadata: [],
+        ctaBanner: CTA_BANNER,
+        perPage: 24,
+        stories: [],
+        total: 0,
+        loadingStories: false
+      };
+    },
+
+    async fetch() {
+      await this.fetchStoryMetadata();
+      this.fetchStories();
+    },
+
+    computed: {
+      selectedTags() {
+        return this.$route.query.tags?.split(',') || [];
+      },
+      filteredTags() {
+        const relevantTags = this.relevantStoryMetadata.map((story) => story.cats.items.filter((cat) => !!cat).map((cat) => cat.id)).flat();
+
+        const tagsSortedByMostUsed = relevantTags.map((tag, i, array) => {
+          return { tag, total: array.filter(t => t === tag).length };
+        }).sort((a, b) => b.total - a.total).map(tag => tag.tag);
+
+        return uniq(tagsSortedByMostUsed);
+      },
+      relevantStoryMetadata() {
+        if (this.selectedTags.length > 0) {
+          // Filter by selected categories
+          return this.allStoryMetadata.filter((story) => {
+            const storyTags = story.cats.items.map((cat) => cat?.id);
+            return this.selectedTags.every((tag) => storyTags.includes(tag));
+          });
+        }
+        return this.allStoryMetadata;
+      },
+      page() {
+        return Number(this.$route.query.page || 1);
+      }
+    },
+
+    watch: {
+      page: 'fetchStories',
+      selectedTags: 'fetchStories'
+    },
+
+
+    methods: {
+      async fetchStoryMetadata() {
+        // Fetch minimal data for all stories to support ordering by datePublished
+        // and filtering by categories.
+        const storyIdsVariables = {
+          locale: this.$i18n.isoLocale(),
+          preview: this.$route.query.mode === 'preview'
+        };
+        const storyIdsResponse = await this.$contentful.query('storiesMinimal', storyIdsVariables);
+        const storyIds = [
+          storyIdsResponse.data.data.blogPostingCollection.items,
+          storyIdsResponse.data.data.exhibitionPageCollection.items
+        ].flat();
+
+        // Order by date published
+        this.allStoryMetadata = storyIds.sort((a, b) => (new Date(b.date)).getTime() - (new Date(a.date)).getTime());
+      },
+
+      async fetchStories() {
+        if (!this.loadingStories) {
+          this.loadingStories = true;
+          // Paginate
+          this.total = this.relevantStoryMetadata.length;
+          const sliceFrom = (this.page - 1) * this.perPage;
+          const sliceTo = sliceFrom + this.perPage;
+          const storySysIds = this.relevantStoryMetadata.slice(sliceFrom, sliceTo).map(story => story.sys.id);
+
+          // Fetch full data for display of page of stories
+          const storiesVariables = {
+            locale: this.$i18n.isoLocale(),
+            preview: this.$route.query.mode === 'preview',
+            limit: this.perPage,
+            ids: storySysIds
+          };
+          const storiesResponse = await this.$contentful.query('storiesBySysId', storiesVariables);
+          const fullStories = [
+            storiesResponse.data.data.blogPostingCollection.items,
+            storiesResponse.data.data.exhibitionPageCollection.items
+          ].flat();
+          this.stories = storySysIds.map((sysId) => fullStories.find((story) => story.sys.id === sysId)).filter(Boolean);
+          if (this.page === 1 && this.selectedTags.length === 0) {
+            this.stories.splice(12, 0, this.ctaBanner);
+          }
+          this.loadingStories = false;
+          this.$scrollTo && this.$scrollTo('#header');
+        }
+      },
+
+      entryUrl(entry) {
+        let urlPrefix;
+
+        if (entry['__typename'] === 'BlogPosting') {
+          urlPrefix = '/blog';
+        } else if (entry['__typename'] === 'ExhibitionPage') {
+          urlPrefix = '/exhibitions';
+        }
+
+        return `${urlPrefix}/${entry.identifier}`;
+      },
+
+      entryImageOptions(image) {
+        return image.width <= image.height ? { width: 660 } : { height: 620 };
+      }
+    }
+  };
+</script>
+
+<style lang="scss" scoped>
+@import '@/assets/scss/variables';
+
+.cta-banner {
+  flex-basis: 100%;
+
+  @media (min-width: $bp-small) {
+    margin-left: $grid-gutter;
+    margin-right: $grid-gutter;
+  }
+}
+</style>
